+++ components/link/basic/src/index.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import Link from 'react-router/lib/Link'
+import { Link } from 'react-router'
 import cx from 'classnames'
 
 class LinkBasic extends Component {